Move ItemContext to react root and fix imports
