+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Move `ItemContext` to react folder.
+
 ## [0.17.0] - 2020-01-29
 ### Added
 - `Advanced configuration` section in the documentation.
