+++ react/Price.tsx
@@ -4,7 +4,7 @@ import { FormattedPrice } from 'vtex.formatted-price'
 import { Loading } from 'vtex.render-runtime'
 import { useCssHandles } from 'vtex.css-handles'
 
-import { useItemContext } from './components/ItemContext'
+import { useItemContext } from './ItemContext'
 import { opaque } from './utils/opaque'
 import { parseTextAlign, TextAlignProp } from './utils/textAlign'
 import styles from './styles.css'
