+++ react/ProductBrand.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent } from 'react'
 import { Loading } from 'vtex.render-runtime'
 import { useCssHandles } from 'vtex.css-handles'
 
-import { useItemContext } from './components/ItemContext'
+import { useItemContext } from './ItemContext'
 import { opaque } from './utils/opaque'
 
 const CSS_HANDLES = ['productBrandName'] as const
