+++ react/QuantitySelector.tsx
@@ -3,7 +3,7 @@ import { Loading } from 'vtex.render-runtime'
 import { useCssHandles } from 'vtex.css-handles'
 
 import Selector from './components/QuantitySelector'
-import { useItemContext } from './components/ItemContext'
+import { useItemContext } from './ItemContext'
 import { AVAILABLE } from './constants/Availability'
 import { opaque } from './utils/opaque'
 import styles from './styles.css'
